Create TEST -Life, the Universe, and Everything
Your program is to use the brute-force approach in order to find the Answer to Life, the Universe, and Everything.More precisely... rewrite small numbers from input to output. Stop processing input after reading in the number 42. All numbers at input are integers of one or two digits.[ExampleInput:
1
2
8
8
4
2
9
9
Output:
1
2
8
8